+++ com.unity.render-pipelines.high-definition/Editor/RenderPipeline/HDEditorUtils.cs
@@ -223,6 +223,54 @@ namespace UnityEditor.Experimental.Rendering.HDPipeline
                 return res.ToString("n2") + " GB";
             }
         }
+
+        /// <summary>Provide a specific property drawer for LightLayer</summary>
+        /// <param name="label">The desired label</param>
+        /// <param name="property">The SerializedProperty (representing an int that should be displayed as a LightLayer)</param>
+        public static void LightLayerMaskPropertyDrawer(GUIContent label, SerializedProperty property)
+        {
+            var renderingLayerMask = property.intValue;
+            int lightLayer;
+            if (property.hasMultipleDifferentValues)
+            {
+                EditorGUI.showMixedValue = true;
+                lightLayer = 0;
+            }
+            else
+                lightLayer = HDAdditionalLightData.RenderingLayerMaskToLightLayer(renderingLayerMask);
+            EditorGUI.BeginChangeCheck();
+            lightLayer = System.Convert.ToInt32(EditorGUILayout.EnumFlagsField(label, (LightLayerEnum)lightLayer));
+            if (EditorGUI.EndChangeCheck())
+            {
+                lightLayer = HDAdditionalLightData.LightLayerToRenderingLayerMask(lightLayer, renderingLayerMask);
+                property.intValue = lightLayer;
+            }
+            EditorGUI.showMixedValue = false;
+        }
+
+        /// <summary>Provide a specific property drawer for LightLayer (without label)</summary>
+        /// <param name="rect">The rect where to draw</param>
+        /// <param name="property">The SerializedProperty (representing an int that should be displayed as a LightLayer)</param>
+        public static void LightLayerMaskPropertyDrawer(Rect rect, SerializedProperty property)
+        {
+            var renderingLayerMask = property.intValue;
+            int lightLayer;
+            if (property.hasMultipleDifferentValues)
+            {
+                EditorGUI.showMixedValue = true;
+                lightLayer = 0;
+            }
+            else
+                lightLayer = HDAdditionalLightData.RenderingLayerMaskToLightLayer(renderingLayerMask);
+            EditorGUI.BeginChangeCheck();
+            lightLayer = System.Convert.ToInt32(EditorGUI.EnumFlagsField(rect, (LightLayerEnum)lightLayer));
+            if (EditorGUI.EndChangeCheck())
+            {
+                lightLayer = HDAdditionalLightData.LightLayerToRenderingLayerMask(lightLayer, renderingLayerMask);
+                property.intValue = lightLayer;
+            }
+            EditorGUI.showMixedValue = false;
+        }
     }
 
     public static partial class SerializedPropertyExtention