+++ com.unity.render-pipelines.high-definition/Runtime/RenderPipeline/RenderPipelineResources.cs
@@ -61,14 +61,16 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
             public ComputeShader buildMaterialFlagsCS;
             [Reload("Runtime/Lighting/LightLoop/Deferred.compute")]
             public ComputeShader deferredCS;
-            [Reload("Runtime/Lighting/Shadow/ScreenSpaceShadow.compute")]
-            public ComputeShader screenSpaceShadowCS;
+            [Reload("Runtime/Lighting/Shadow/ContactShadows.compute")]
+            public ComputeShader contactShadowCS;
             [Reload("Runtime/Lighting/VolumetricLighting/VolumeVoxelization.compute")]
             public ComputeShader volumeVoxelizationCS;
             [Reload("Runtime/Lighting/VolumetricLighting/VolumetricLighting.compute")]
             public ComputeShader volumetricLightingCS;
             [Reload("Runtime/Lighting/LightLoop/DeferredTile.shader")]
             public Shader deferredTilePS;
+            [Reload("Runtime/Lighting/Shadow/ScreenSpaceShadows.shader")]
+            public Shader screenSpaceShadowPS;
 
             [Reload("Runtime/Material/SubsurfaceScattering/SubsurfaceScattering.compute")]
             public ComputeShader subsurfaceScatteringCS;                // Disney SSS
@@ -303,7 +305,9 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
             public Texture2D debugFontTex;
             [Reload("Runtime/Debug/ColorGradient.png")]
             public Texture2D colorGradient;
-            
+            [Reload("Runtime/RenderPipelineResources/Texture/Matcap/DefaultMatcap.png")]
+            public Texture2D matcapTex;
+
             // Pre-baked noise
             [Reload("Runtime/RenderPipelineResources/Texture/BlueNoise16/L/LDR_LLL1_{0}.png", 0, 32)]
             public Texture2D[] blueNoise16LTex;