+++ com.unity.render-pipelines.lightweight/Editor/2D/ShapeEditor/GUIFramework/HoveredControlAction.cs
@@ -1,7 +1,7 @@
 using System;
 using UnityEngine;
 
-namespace UnityEditor.Experimental.Rendering.LWRP.GUIFramework
+namespace UnityEditor.Experimental.Rendering.LWRP.Path2D.GUIFramework
 {
     internal abstract class HoveredControlAction : GUIAction
     {